refactor: cleanup and typofixes
Relates #256
