+++ app/(dashboard)/search/page.tsx
@@ -1,13 +1,11 @@
-import { SearchBar } from '@/components/SearchBar';
+import { SearchBar } from '@/components/filter/SearchBar';
 import QuickBrowse from '@/components/filter/QuickBrowse';
 
 export default function SearchItemPage() {
   return (
     <div className='mt-8'>
       <SearchBar />
-      <div className='mt-4'>
-        <QuickBrowse />
-      </div>
+      <QuickBrowse />
     </div>
   );
 }
